browser(firefox-stable): roll Firefox to 104.0
References #16780
