+++ browser_patches/firefox/BUILD_NUMBER
@@ -1,2 +1,2 @@
-1346
-Changed: yurys@chromium.org Mon Aug 22 16:56:24 PDT 2022
+1347
+Changed: lushnikov@chromium.org Tue 23 Aug 2022 03:05:32 PM PDT
