+++ browser_patches/firefox/UPSTREAM_CONFIG.sh
@@ -1,3 +1,3 @@
 REMOTE_URL="https://github.com/mozilla/gecko-dev"
 BASE_BRANCH="release"
-BASE_REVISION="f532fa26bc150add9f8ba42e5fbfdbeffa23662f"
+BASE_REVISION="fd854580ffc6fba6a0acdf335c96a1b24b976cb9"
