+++ browser_patches/firefox/juggler/components/moz.build
@@ -2,8 +2,5 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-EXTRA_COMPONENTS += [
-    "juggler.js",
-    "juggler.manifest",
-]
+XPCOM_MANIFESTS += ["components.conf"]
